requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.7.0
flake8==7.1.2
frozenlist==1.5.0
idna==3.10
isort==6.0.1
lxml==5.3.1
mccabe==0.7.0
motor==3.7.0
multidict==6.1.0
propcache==0.3.0
pycodestyle==2.12.1
pyflakes==3.2.0
pymongo==4.11.2
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
